Extract parser abstract class from Android RSS parser
